@@ -1,26 +1,34 @@
 package me.kodysimpson.chunkcollector;
 
-import me.kodysimpson.chunkcollector.commands.CommandManager;
+import me.kodysimpson.chunkcollector.commands.subcommands.BuyCommand;
+import me.kodysimpson.chunkcollector.commands.subcommands.GiveCommand;
+import me.kodysimpson.chunkcollector.commands.subcommands.ReloadCommand;
+import me.kodysimpson.chunkcollector.database.Database;
 import me.kodysimpson.chunkcollector.listeners.CollectorListener;
-import me.kodysimpson.chunkcollector.menusystem.PlayerMenuUtility;
-import me.kodysimpson.chunkcollector.tasks.CollectDrops;
-import me.kodysimpson.chunkcollector.utils.Database;
+import me.kodysimpson.chunkcollector.tasks.AutoKill;
+import me.kodysimpson.chunkcollector.tasks.ChunkScanner;
+import me.kodysimpson.chunkcollector.tasks.MineTask;
+import me.kodysimpson.simpapi.command.CommandList;
+import me.kodysimpson.simpapi.command.CommandManager;
+import me.kodysimpson.simpapi.command.SubCommand;
+import me.kodysimpson.simpapi.menu.MenuManager;
 import net.milkbowl.vault.economy.Economy;
 import org.bukkit.Bukkit;
+import org.bukkit.ChatColor;
 import org.bukkit.block.Block;
-import org.bukkit.entity.Player;
+import org.bukkit.command.CommandSender;
 import org.bukkit.plugin.RegisteredServiceProvider;
 import org.bukkit.plugin.java.JavaPlugin;
-import org.bukkit.scheduler.BukkitTask;
 
+import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.logging.Logger;
 
 public final class ChunkCollector extends JavaPlugin {
 
     private static final Logger log = Logger.getLogger("Minecraft");
-    public static HashMap<Player, PlayerMenuUtility> playerMenuUtilityMap = new HashMap<>();
     //collector ID - fully grown crops
     public static HashMap<Integer, ArrayList<Block>> crops = new HashMap<>();
     private static Economy econ = null;
@@ -32,25 +40,7 @@ public final class ChunkCollector extends JavaPlugin {
     }
 
     public static String getConnectionURL() {
-
         return url;
-
-    }
-
-    //Provide a player and return a menu system for that player
-    //create one if they don't already have one
-    public static PlayerMenuUtility getPlayerMenuUtility(Player p) {
-        PlayerMenuUtility playerMenuUtility;
-        if (!(playerMenuUtilityMap.containsKey(p))) { //See if the player has a lockmenusystem "saved" for them
-
-            //This player doesn't. Make one for them add add it to the hashmap
-            playerMenuUtility = new PlayerMenuUtility(p);
-            playerMenuUtilityMap.put(p, playerMenuUtility);
-
-            return playerMenuUtility;
-        } else {
-            return playerMenuUtilityMap.get(p); //Return the object by using the provided player
-        }
     }
 
     public static Economy getEconomy() {
@@ -83,19 +73,52 @@ public final class ChunkCollector extends JavaPlugin {
         url = "jdbc:h2:" + getDataFolder().getAbsolutePath() + "/data/chunkcollector";
         Database.initializeDatabase();
 
-        //Command manager
-        getCommand("collector").setExecutor(new CommandManager());
+        try {
+            CommandManager.createCoreCommand(this, "collector", "Use the chunk collector plugin", "/collector", new CommandList() {
+                @Override
+                public void displayCommandList(CommandSender p, List<SubCommand> subCommandList) {
+                    p.sendMessage(" ");
+                    p.sendMessage(ChatColor.GREEN + "======= " + ChatColor.GRAY + "[" + ChatColor.LIGHT_PURPLE + ChatColor.BOLD + "Chunk" + ChatColor.AQUA + "Collector" + ChatColor.GRAY + "] " + ChatColor.YELLOW + "Commands " + ChatColor.GREEN + "=======");
+                    p.sendMessage(" ");
+                    for (SubCommand subCommand : subCommandList) {
+                        p.sendMessage(ChatColor.DARK_GRAY + " - " + ChatColor.YELLOW + subCommand.getSyntax() + " - " + ChatColor.GRAY + subCommand.getDescription());
+                    }
+                    p.sendMessage(" ");
+                    p.sendMessage(ChatColor.GREEN + "=====================================");
+                    p.sendMessage(" ");
+                }
+            }, List.of("cc", "chunkcollector"), BuyCommand.class, GiveCommand.class, ReloadCommand.class);
+        } catch (NoSuchFieldException | IllegalAccessException e) {
+            e.printStackTrace();
+        }
+
+        MenuManager.setup(getServer(), this);
 
         //Listeners
         Bukkit.getServer().getPluginManager().registerEvents(new CollectorListener(), this);
 
         //Collection task[starts up 1 minute from startup]
-        BukkitTask task = new CollectDrops().runTaskTimer(this, 1200, getConfig().getLong("collection-duration"));
+        new ChunkScanner().runTaskTimer(this, 1200, getConfig().getLong("collection-duration"));
+
+        if (getConfig().getBoolean("auto-kill")) {
+            new AutoKill().runTaskTimer(this, 1200, getConfig().getLong("auto-kill-interval"));
+        }
+
+        new MineTask().runTaskTimer(this, 600, 1200);
 
     }
 
     @Override
     public void onDisable() {
+
+        //Close the current connection to the DB
+        try {
+            System.out.println("CLOSING THE DATABASE CONNECTION");
+            Database.getConnection().close();
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+
         // Plugin shutdown logic
         log.info(String.format("[%s] Disabled Version %s", getDescription().getName(), getDescription().getVersion()));
     }
