@@ -1,33 +1,53 @@
 package me.kodysimpson.chunkcollector.menusystem.menus;
 
 import me.kodysimpson.chunkcollector.ChunkCollector;
-import me.kodysimpson.chunkcollector.config.Config;
-import me.kodysimpson.chunkcollector.menusystem.PaginatedMenu;
-import me.kodysimpson.chunkcollector.menusystem.PlayerMenuUtility;
-import me.kodysimpson.chunkcollector.utils.Collector;
-import me.kodysimpson.chunkcollector.utils.Database;
+import me.kodysimpson.chunkcollector.database.Database;
+import me.kodysimpson.chunkcollector.model.Collector;
 import me.kodysimpson.chunkcollector.utils.Utils;
+import me.kodysimpson.simpapi.colors.ColorTranslator;
+import me.kodysimpson.simpapi.exceptions.MenuManagerException;
+import me.kodysimpson.simpapi.exceptions.MenuManagerNotSetupException;
+import me.kodysimpson.simpapi.menu.PaginatedMenu;
+import me.kodysimpson.simpapi.menu.PlayerMenuUtility;
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
 import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.inventory.ItemStack;
+import org.jetbrains.annotations.Nullable;
 
-import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
 
 public class CollectorStorageMenu extends PaginatedMenu {
 
+    private final Collector collector;
+
     public CollectorStorageMenu(PlayerMenuUtility playerMenuUtility) {
         super(playerMenuUtility);
+
+        collector = Database.getCollectorDataAccess().findById(playerMenuUtility.getData(MenuData.COLLECTOR_ID, Integer.class));
+    }
+
+    @Override
+    public <T> List<T> getData() {
+        return (List<T>) Utils.combine(collector.getItems());
+    }
+
+    @Override
+    public void loopCode(Object object) {
+        inventory.addItem((ItemStack) object);
     }
 
     @Override
     public String getMenuName() {
-        if (playerMenuUtility.getType() == Database.CollectionType.DROP) {
-            return ChatColor.translateAlternateColorCodes('&', ChunkCollector.getPlugin().getConfig().getString("Menu Titles.Collector-Storage Menu.drop"));
-        } else {
-            return ChatColor.translateAlternateColorCodes('&', ChunkCollector.getPlugin().getConfig().getString("Menu Titles.Collector-Storage Menu.crop"));
-        }
+
+        return switch (collector.getType()) {
+            case DROP -> ColorTranslator.translateColorCodes(ChunkCollector.getPlugin().getConfig().getString("Menu Titles.Collector-Storage Menu.drop"));
+            case CROP -> ColorTranslator.translateColorCodes(ChunkCollector.getPlugin().getConfig().getString("Menu Titles.Collector-Storage Menu.crop"));
+            case ORE -> ColorTranslator.translateColorCodes(ChunkCollector.getPlugin().getConfig().getString("Menu Titles.Collector-Storage Menu.ore"));
+        };
+
     }
 
     @Override
@@ -36,12 +56,24 @@ public class CollectorStorageMenu extends PaginatedMenu {
     }
 
     @Override
-    public void handleMenu(InventoryClickEvent e) {
+    public boolean cancelAllClicks() {
+        return true;
+    }
+
+    @Override
+    public @Nullable HashMap<Integer, ItemStack> getCustomMenuBorderItems() {
+
+        HashMap<Integer, ItemStack> borderItems = new HashMap<>();
+        ItemStack sellAllItems = makeItem(Material.FIREWORK_ROCKET, ColorTranslator.translateColorCodes("&e&lSell All Items"));
+        borderItems.put(46, sellAllItems);
+
+        return borderItems;
+    }
+
+    @Override
+    public void handleMenu(InventoryClickEvent e) throws MenuManagerException, MenuManagerNotSetupException {
+
         Player p = (Player) e.getWhoClicked();
-
-        Collector collector = Database.findByID(playerMenuUtility.getCollectorID());
-
-        ArrayList<ItemStack> drops = Utils.combine(collector.getItems());
 
         if (e.getCurrentItem().getType() == Material.BARRIER) {
 
@@ -52,46 +84,21 @@ public class CollectorStorageMenu extends PaginatedMenu {
             Utils.sellAllItems(collector);
 
             //Reload the page
-            super.open();
+            reload();
 
         } else if (e.getCurrentItem().getType() == Material.DARK_OAK_BUTTON) {
+
             if (ChatColor.stripColor(e.getCurrentItem().getItemMeta().getDisplayName()).equalsIgnoreCase("Left")) {
-                if (page == 0) {
-                    p.sendMessage(Config.FIRST_PAGE);
-                } else {
-                    page = page - 1;
-                    super.open();
+
+                if (!prevPage()) {
+                    p.sendMessage(ColorTranslator.translateColorCodes(ChunkCollector.getPlugin().getConfig().getString("Messages.first-page")));
                 }
             } else if (ChatColor.stripColor(e.getCurrentItem().getItemMeta().getDisplayName()).equalsIgnoreCase("Right")) {
-                if (!((index + 1) >= drops.size())) {
-                    page = page + 1;
-                    super.open();
-                } else {
-                    p.sendMessage(Config.LAST_PAGE);
+                if (!nextPage()) {
+                    p.sendMessage(ColorTranslator.translateColorCodes(ChunkCollector.getPlugin().getConfig().getString("Messages.last-page")));
                 }
             }
         }
-    }
-
-    @Override
-    public void setMenuItems() {
-
-        addMenuBorder();
-
-        Collector collector = Database.findByID(playerMenuUtility.getCollectorID());
-
-        ArrayList<ItemStack> drops = Utils.combine(collector.getItems());
-
-        if (drops != null && !drops.isEmpty()) {
-            for (int i = 0; i < getMaxItemsPerPage(); i++) {
-                index = getMaxItemsPerPage() * page + i;
-                if (index >= drops.size()) break;
-                if (drops.get(index) != null) {
-                    inventory.addItem(drops.get(index));
-                }
-            }
-        }
-
-
     }
 }
+
