@@ -1,10 +1,11 @@
 package me.kodysimpson.chunkcollector.menusystem.menus;
 
 import me.kodysimpson.chunkcollector.ChunkCollector;
-import me.kodysimpson.chunkcollector.menusystem.Menu;
-import me.kodysimpson.chunkcollector.menusystem.PlayerMenuUtility;
-import me.kodysimpson.chunkcollector.utils.Database;
+import me.kodysimpson.chunkcollector.utils.CollectionType;
 import me.kodysimpson.chunkcollector.utils.Utils;
+import me.kodysimpson.simpapi.colors.ColorTranslator;
+import me.kodysimpson.simpapi.menu.Menu;
+import me.kodysimpson.simpapi.menu.PlayerMenuUtility;
 import net.milkbowl.vault.economy.EconomyResponse;
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
@@ -14,17 +15,23 @@ import org.bukkit.inventory.ItemStack;
 
 public class ConfirmBuyMenu extends Menu {
 
+    private final CollectionType buyType;
+
     public ConfirmBuyMenu(PlayerMenuUtility playerMenuUtility) {
         super(playerMenuUtility);
+        this.buyType = playerMenuUtility.getData(MenuData.BUY_TYPE, CollectionType.class);
     }
 
     @Override
     public String getMenuName() {
-        if (playerMenuUtility.getBuyType() == Database.CollectionType.DROP){
-            return ChatColor.translateAlternateColorCodes('&', ChunkCollector.getPlugin().getConfig().getString("Menu Titles.Confirm-Buy Menu.drop"));
-        }else{
-            return ChatColor.translateAlternateColorCodes('&', ChunkCollector.getPlugin().getConfig().getString("Menu Titles.Confirm-Buy Menu.crop"));
+        if (buyType == CollectionType.DROP) {
+            return ColorTranslator.translateColorCodes(ChunkCollector.getPlugin().getConfig().getString("Menu Titles.Confirm-Buy Menu.drop"));
+        } else if (buyType == CollectionType.CROP) {
+            return ColorTranslator.translateColorCodes(ChunkCollector.getPlugin().getConfig().getString("Menu Titles.Confirm-Buy Menu.crop"));
+        } else if (buyType == CollectionType.ORE) {
+            return ColorTranslator.translateColorCodes(ChunkCollector.getPlugin().getConfig().getString("Menu Titles.Confirm-Buy Menu.ore"));
         }
+        return "pickle";
     }
 
     @Override
@@ -33,19 +40,26 @@ public class ConfirmBuyMenu extends Menu {
     }
 
     @Override
+    public boolean cancelAllClicks() {
+        return true;
+    }
+
+    @Override
     public void handleMenu(InventoryClickEvent e) {
 
-        ItemStack collector;
+        ItemStack collector = null;
         Player p = playerMenuUtility.getOwner();
 
-        if ((e.getCurrentItem().getType() == Material.DIAMOND_SWORD) || (e.getCurrentItem().getType() == Material.BREAD)){
+        if ((e.getCurrentItem().getType() == Material.DIAMOND_SWORD) || (e.getCurrentItem().getType() == Material.BREAD) || (e.getCurrentItem().getType() == Material.GOLD_ORE)) {
 
             double cost = 0;
 
-            if (playerMenuUtility.getBuyType() == Database.CollectionType.DROP){
+            if (buyType == CollectionType.DROP) {
                 cost = ChunkCollector.getPlugin().getConfig().getDouble("collector-cost.drop");
-            }else{
+            } else if (buyType == CollectionType.CROP) {
                 cost = ChunkCollector.getPlugin().getConfig().getDouble("collector-cost.crop");
+            } else if (buyType == CollectionType.ORE) {
+                cost = ChunkCollector.getPlugin().getConfig().getDouble("collector-cost.ore");
             }
 
             //see if they have enough
@@ -55,12 +69,15 @@ public class ConfirmBuyMenu extends Menu {
 
                 if (response.transactionSuccess()){
 
-                    if (playerMenuUtility.getBuyType() == Database.CollectionType.DROP){
+                    if (buyType == CollectionType.DROP) {
                         p.sendMessage(ChatColor.GREEN + "Purchased Drop Collector for $" + ChatColor.YELLOW + cost);
-                        collector = Utils.makeCollector(p, Database.CollectionType.DROP);
-                    }else{
+                        collector = Utils.makeCollector(p, CollectionType.DROP);
+                    } else if (buyType == CollectionType.CROP) {
                         p.sendMessage(ChatColor.GREEN + "Purchased Crop Collector for $" + ChatColor.YELLOW + cost);
-                        collector = Utils.makeCollector(p, Database.CollectionType.CROP);
+                        collector = Utils.makeCollector(p, CollectionType.CROP);
+                    } else if (buyType == CollectionType.ORE) {
+                        p.sendMessage(ChatColor.GREEN + "Purchased Ore Collector for $" + ChatColor.YELLOW + cost);
+                        collector = Utils.makeCollector(p, CollectionType.ORE);
                     }
 
                     if (p.getInventory().firstEmpty() == -1){
@@ -93,17 +110,23 @@ public class ConfirmBuyMenu extends Menu {
     @Override
     public void setMenuItems() {
 
-        if (playerMenuUtility.getBuyType() == Database.CollectionType.DROP){
+        if (buyType == CollectionType.DROP) {
 
             inventory.setItem(3, makeItem(Material.DIAMOND_SWORD, ChatColor.GREEN + "" + ChatColor.BOLD + "Purchase",
                     ChatColor.AQUA + "Buy Drop Collector for",
                     ChatColor.GOLD + "$" + ChunkCollector.getPlugin().getConfig().getDouble("collector-cost.drop")));
 
-        }else if(playerMenuUtility.getBuyType() == Database.CollectionType.CROP){
+        } else if (buyType == CollectionType.CROP) {
 
             inventory.setItem(3, makeItem(Material.BREAD, ChatColor.GREEN + "" + ChatColor.BOLD + "Purchase",
                     ChatColor.AQUA + "Buy Crop Collector for",
                     ChatColor.GOLD + "$" + ChunkCollector.getPlugin().getConfig().getDouble("collector-cost.crop")));
+
+        } else if (buyType == CollectionType.ORE) {
+
+            inventory.setItem(3, makeItem(Material.BREAD, ChatColor.GREEN + "" + ChatColor.BOLD + "Purchase",
+                    ChatColor.AQUA + "Buy Ore Collector for",
+                    ChatColor.GOLD + "$" + ChunkCollector.getPlugin().getConfig().getDouble("collector-cost.ore")));
 
         }
 
